Move retrying to API call
