+++ tests/darwin/dataset/remote_dataset_test.py
@@ -9,8 +9,11 @@ import pytest
 import requests
 import responses
 from pydantic import ValidationError
-from requests.exceptions import RequestException
+from requests.exceptions import HTTPError, RequestException
+from requests.models import Response
+from tenacity import RetryCallState, Retrying
 
+from darwin.backend_v2 import BackendV2, retry_if_status_code_429
 from darwin.client import Client
 from darwin.config import Config
 from darwin.dataset import RemoteDataset
@@ -1079,42 +1082,3 @@ class TestRegisterMultiSlotted:
         )
         assert len(result["registered"]) == 0
         assert len(result["blocked"]) == 1
-
-
-@pytest.mark.usefixtures("file_read_write_test")
-class TestRegisterItemsWithRetry:
-    @responses.activate
-    def test_register_items_with_retry_success(self, remote_dataset: RemoteDatasetV2):
-        responses.add(
-            responses.POST,
-            "http://localhost/api/v2/teams/test_team/items/register_existing",
-            json={
-                "items": [{"id": "1", "name": "test.jpg"}],
-                "blocked_items": [],
-            },
-            status=200,
-        )
-        response = remote_dataset.register_items_with_retry(
-            {"item1": ["test.jpg"]}, "test_team"
-        )
-        assert (
-            response.text
-            == '{"items": [{"id": "1", "name": "test.jpg"}], "blocked_items": []}'
-        )
-        assert response.status_code == 200
-
-    @responses.activate
-    def test_register_items_with_retry_throws_exception(
-        self, remote_dataset: RemoteDatasetV2
-    ):
-        with patch.object(
-            remote_dataset,
-            "register_items_with_retry",
-            side_effect=requests.exceptions.RequestException(
-                "HTTP 429: Too Many Requests"
-            ),
-        ):
-            with pytest.raises(requests.exceptions.RequestException):
-                remote_dataset.register_items_with_retry(
-                    {"item1": ["test.jpg"]}, "test_team"
-                )
